Renamed EUnicodeBidi to just UnicodeBidi
Renamed EUnicodeBidi to just UnicodeBidi to match the name of the file,
UnicodeBidi.h. This is pre-work to allow UnicodeBidi to be used by
generated fields in ComputedStyleBase, and also better matches the
style guidelines (class names should have the same name as the file).

BUG=628043

Review-Url: https://codereview.chromium.org/2620873002
Cr-Original-Commit-Position: refs/heads/master@{#443041}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bcbe36cce42ae13bee7b03a6e037237676959935