Don't force-update style and layout in Document::scrollingElement
This method has callsites inside of compositing, in which it's illegal
to try to update layout and style because compositing comes after.
However, it's also pointless to do so from such callsites.

The other call sites are: FrameView::scrollBehaviorStyle
Various scrolling methods in Element. All of the Element callsites already
update style and layout; this CL adds it to the former (which has a very
complicated set of ways it can be called, making it very hard to determine
if style or layout could be dirty before calling it).

BUG=679099
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2618323004
Cr-Original-Commit-Position: refs/heads/master@{#443049}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e8c827b638b54a43e065cd6b3a431fefc5ae9768