Extract diff_text function out of the Port class.
This is a refactoring change related to, but separate from,
the issue of replacing wdiff and PrettyPatch.rb with a Python
diff module (http://crbug.com/672651), and a follow-up to
adding html_diff.py (http://crrev.com/2580143002).

Reason: The functionality of getting a text diff shouldn't be platform
dependent, so it doesn't really seem to belong in the Port class. (It
seems more similar to the html_diff module in webkitpy/common/.)

Review-Url: https://codereview.chromium.org/2595983002
Cr-Original-Commit-Position: refs/heads/master@{#443067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 938778c92da3d97c66a4fa16c0d79a019c0c40da