Generalize browser controls adjustment for arbitrary scrollers.
When the browser controls (URL bar) are hidden/shown, they change the visible
area of the viewport, allowing more or less scrolling. Until a resize occurs and
the FrameView/LayoutView is resized, we keep an "adjustment" value that we use to
fudge the maximumScrollOffset in the interim.

When a scroller is set as the document.rootScroller, we want it to behave in
the same way. This patch extends the behavior to all FrameViews and
PaintLayerScrollableAreas when they become the global root scroller. The
adjustment value is no longer stored in the FrameView. Instead,
TopDocumentRootScrollerController uses the value stored in VisualViewport to
calculate the current visible area of the viewport.

BUG=505516
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2622103002
Cr-Original-Commit-Position: refs/heads/master@{#443079}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eea21c858d184a593903840cbaaa08339f6ac603