Reset ResourceLoader::m_loader in the pre-finalizer
A ResourceLoader is held as a raw-ptr in WebURLLoaderImpl and hence it must
cancel the WebURLLoaderImpl when it's destroyed. We thought
blink::Resource::allClientsAndObserversRemoved did the job but it's wrong:
the cancel timer doesn't work when the resource is about to be destroyed.

BUG=675947
R=hiroshige@chromium.org, japhet@chromium.org

Review-Url: https://codereview.chromium.org/2625823003
Cr-Original-Commit-Position: refs/heads/master@{#443167}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e6d7a20f7cb4b23ff305b16579faf94f7ddf992