Put an early return if an interface has a runtime enabled flag and it is false.
It was confusing that the flag seemed to check if enable some
specific members (attributes, methods, and constants), but it
should check everything of the interface.
It means if the flag is false, we don't have to check all
members, e.g. indexed properties and iterators.

BUG=None

Review-Url: https://codereview.chromium.org/2620153004
Cr-Original-Commit-Position: refs/heads/master@{#443179}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c162e0c69d45b4bb191cb80ca9fde153950a270b