Introduce DOMWindowClient
The lifetime of ExecutionContext and DOMWindow is a bit different.
In common case, they get detached at the same timing but in a case where
a frame navigates from an initial empty document to another same-origin document,
the ExecutionContext gets detached but the DOMWindow is reused.
Due to the (subtle) difference, we need DOMWindowClient in addition to ContextClient.

ContextClient is a client for objects that want to observe ExecutionContext.
DOMWindowClient is a client for objects that want to observe DOMWindow.

This CL uses DOMWindowClient for Navigator.

BUG=610176

Review-Url: https://codereview.chromium.org/2629493004
Cr-Original-Commit-Position: refs/heads/master@{#443216}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a82bb82ae350134c763a8e995738fec5d4e2a061