Fixed text-align keywords in CSSProperties.in.
Fixed text-align keywords in CSSProperties.in to match the values in
CSSValueKeywords.in. This will not actually change the generated values
at this time since "webkitLeft", and "-webkit-left" will both result
in the enum value "kWebkitLeft" being generated. However, long term
it is important that the values exactly match those in
CSSValueKeywords.in.

BUG=628043

Review-Url: https://codereview.chromium.org/2631473002
Cr-Original-Commit-Position: refs/heads/master@{#443410}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5db4d106559db54bf67baf7dc6cce91b41dca812