Reduce rectangle merges when doing region-based hit test on text boxes
Only unite the box rect with the hit test result if it intersects the hit
test rect. That is sufficient because the only use for the united rect
is to check that it contains the location rect for early termination of
testing.

Patch based on work by oicqchen@gmail.com.

R=wkorman@chromium.org
BUG=676764

Review-Url: https://codereview.chromium.org/2616483002
Cr-Original-Commit-Position: refs/heads/master@{#443684}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f24093ff9c2a18dcddb8728edf40be941b8aadc