Make CSSParserContext be garbage collected.
This is a refactor only, no functionality should change. This allocates
CSSParserContext on the heap and wires it up to be garbage collected.
This is necessary so in the future it can have a handle to a Document
(also a GC'ed class) instead of a UseCounter.

Also includes some minor API changes to CSSParserContext that I think
make it more readable, since I'm changing every callsite of the class
already anyway.

TBR=haraken@chromium.org
BUG=668251

Review-Url: https://codereview.chromium.org/2616093003
Cr-Original-Commit-Position: refs/heads/master@{#443792}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 37b65afaf587c59143a6dca1ea75a505d6c696c6