Enable ServiceWorkerNavigationPreload by default for Origin-Trial.
The CL https://codereview.chromium.org/2627023002/ introduced the mechanism for
Origin-Trial experiment of Service Worker Navigation Preload. But base::Feature
of the feature was disabled by default. So when we start the Origin-Trial
experiment, we had to enable the feature by the field trial mechanism. But it
is possible that the feature is not correctly enabled by the field trial
mechanism (Ex: in the first boot of Chrome). In order to avoid the confusion,
this CL enables the base::Feature of the feature by default.

If something goes wrong and we will want to stop the origin-trial, we can stop
the experiment by disabling this feature using the field trial mechanism.

  * Origin Trial: Have an effective token.
                               Command line
                           Default  Enable  Disabled
                 Default     *A*      A        B
    Field trial  Enabled      A       A        B
                 Disabled     B       A        B

  * Origin Trial: No token.
                               Command line
                           Default  Enable  Disabled
                 Default      C       A        C
    Field trial  Enabled      C       A        C
                 Disabled     C       A        C

  - A: Navigation Preload related methods and attributes are available in JS
      and work correctly.
  - B: Navigation Preload related methods and attributes are available in
      JS. But NavigationPreloadManager's enable, disable and setHeaderValue
      methods always return a rejected promise. And FetchEvent's
      preloadResponse attribute returns a promise which always resolve with
      undefined.
  - C: Navigation Preload related methods and attributes are not available
      in JS.

BUG=649558

Review-Url: https://codereview.chromium.org/2637763002
Cr-Original-Commit-Position: refs/heads/master@{#443851}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9ad964459156c4bd27fa68a0bec4155409ec492d