Complete usage tracking for OffscreenCanvas.convertToBlob API
We've decided to merge the usage data from OffscreenCanvas.convertToBlob to
canvas.toBlob because they share the same logic in CanvasAsyncBlobCreator.
Moreover, previously we forgot to add a new histogram
'Canvas.RequestedImageMimeTypes_convertToBlobPromise' which is recorded in
ImageEncoderUtils::toEncodingMimeType; so in this CL I'm adding it.

BUG=653599

Review-Url: https://codereview.chromium.org/2629923003
Cr-Original-Commit-Position: refs/heads/master@{#443957}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9c7d5f102a02c3cb9e740deab5032c8c27a3a5c6