Remove webkitpy/w3c/wpt_bridge_bot.py.
Reason: This module was intended to be put here temporarily so that parts
could be re-used in the exporter.  Now that the exporter has been added,
I think that this can be removed, and it's still here in version control
if we want to reference it later.

Review-Url: https://codereview.chromium.org/2630213002
Cr-Original-Commit-Position: refs/heads/master@{#444088}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e841b7222f6cb368a8640ac2aaeaf53c7afeac18