Reset DragState::m_dragSrc when mouse move doesn't meet drag threshold.
The rest of blink assumes that the DragState is completely valid if
DragState::m_dragSrc is not null. However, MouseEventManger::handleDrag
has a code path that sets DragState::m_dragSrc, but does not call
tryStartDrag(), which sets DragState::m_dataTransfer. This creates the
opportunity for crashes, as DragState::m_dataTransfer is initialized to
null when a renderer is created.

BUG=677916

Review-Url: https://codereview.chromium.org/2610023002
Cr-Original-Commit-Position: refs/heads/master@{#444252}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 441df0543f6f4029f73068dec974d4dac47f6a1f