Add code path for registered custom properties in CSSInterpolationType
This change adds extra handling logic for registered custom properties
in CSSInterpolationType that deals with their associated CSSValues instead
of just the generic CSSCustomPropertyDeclaration values.

This is in preparation for integration with the typed subclasses
of CSSInterpolationType to produce values in between the specified
animation keyframes.

This patch is a refactor that has no change in behaviour.

BUG=671904

Review-Url: https://codereview.chromium.org/2630503002
Cr-Original-Commit-Position: refs/heads/master@{#444271}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 271557f792311061db727bd05dcedd282c1c2272