Split svg/dom/transform-parser.html
This CL splits the mentioned test into two to alleviate the potential
for test timeouts.
The bulk of the test is moved to a JS file, and the js-test dependency
is dropped since the only useful output of this test is the console log
generated.
The TestExpectations entry is dropped.

BUG=664849

Review-Url: https://codereview.chromium.org/2646503002
Cr-Original-Commit-Position: refs/heads/master@{#444418}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e800d7672ea681b08ecad44d273d3db1689af1c