Percent-width blocks cannot form a re-layout boundary.
A block with non-visible overflow can only form a re-layout boundary if both
width and height are fixed.

The block may be inside a shrink-to-fit container (there's no cheap and
reliable way to detect that), so that changes inside it may affect its width.

BUG=681657

Review-Url: https://codereview.chromium.org/2643703002
Cr-Original-Commit-Position: refs/heads/master@{#444467}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da179e152dff0ab3e3dabc72299bd4fb586881c9