Add PREVIEWS_NO_TRANSFORM bit to PreviewsState for Previews reloads
When the user asks for the original resource from a Preview, the Chrome
Proxy Accept Transform header should not be added to requests since the
server should not be allowed to do transformations on the request. Add
a bit to PreviewsState to indicate that the request is for the original
resource with no transform.

BUG=681097
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2625413005
Cr-Original-Commit-Position: refs/heads/master@{#444545}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8b3b3737eea6c2c42931becd0873110d5fde45cf