Use Unthrottled instead of DOMManipulation in HTMLLinkElement/HTMLStyleElement
The async tasks for HTMLLinkElement/HTMLStyleElement::dispatchPendingEvent()
are on the critical paths to show popup windows for extensions, but
the DOMManipulation task runner is throttled because the tasks are executed
before the popup windows become visible, causing delay in showing popups.

The scheduler is fixed by [1], but [1] is not merged to M-56.
This CL makes a mergeable workaround by using TaskType::Unthrottled for these
tasks to fix the issue on M-56.

[1] https://codereview.chromium.org/2620743002

BUG=649942

Review-Url: https://codereview.chromium.org/2596343002
Cr-Original-Commit-Position: refs/heads/master@{#444559}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 812f90689e6637a5a4f13151dba661464716e29e