Revert 'Make "compositionend" event fired after setting caret position'
This reverts http://crrev.com/2493703002, which was bisected to cause Google
Docs Korean IME issue 674295.  Note the revert is unclean in that the
structural changes to InputMethodControllerTest are preserved (only the
new tests are deleted) and a new updateStyleAndLayoutIgnorePendingStylesheets
was needed because of changes since the original landing.

TBR=yosin@chromium.org
BUG=674295,663944

Review-Url: https://codereview.chromium.org/2644623003
Cr-Original-Commit-Position: refs/heads/master@{#444593}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53003b5bb24d266ac78adf0623b2dd07c1ac6e9f