Implements CSSPropertyAPI for the outline-color property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyOutlineColor from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
outline-color property is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIOutlineColor.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIOutlineColor.h is generated.
- Moves the parsing logic for outline-color from CSSPropertyParser.cpp to
  CSSPropertyAPIOutlineColor.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2639143004
Cr-Original-Commit-Position: refs/heads/master@{#444613}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 381fe1bb9f6bf01ff4c0690ced1143f17f12a805