Implements CSSPropertyAPI for the font-variant-ligatures property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyFontVariantLigatures from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
font-variant-ligatures property is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIFontVariantLigatures.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIFontVariantLigatures.h is generated.
- Moves the parsing logic for font-variant-ligatures from CSSPropertyParser.cpp to
  CSSPropertyAPIFontVariantLigatures.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2645773002
Cr-Original-Commit-Position: refs/heads/master@{#444650}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 886e2c6d39d318467d980657f9d598531f824ac9