Implements CSSPropertyAPI for letter and word spacing
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyLetterSpacing, and CSSPropertyWordSpacing from the switch
statement in parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for
the letter and word spacing properties is stored in a
CSSPropertyDescriptor, and is called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPILetterAndWordSpacing.cpp to the BUILD.gn file.
- Adds api_class=CSSPropertyAPILetterAndWordSpacing flag to each border
  class property in CSSProperties.in, which indicates that
  CSSPropertyAPILetterAndWordSpacing.h is generated.
- Moves the parsing logic for letter and word spacing from
  CSSPropertyParser.cpp to CSSPropertyAPILetterAndWordSpacing.cpp,
  which implements CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2641083002
Cr-Original-Commit-Position: refs/heads/master@{#444675}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0006a31fd4a9fcf1b39a099f9a3056856fa8e063