PlzNavigate: report the accurate total size for the main resource
This CL fixes a discrepency in the reported size of the total transfer
size of teh main resource when PlzNavigate is enabled. This is due to
the response being switched from the one sent by the
AsyncResourceHandler (which corresponds to the requets of the blob url)
to the actual one received when the network request was made. This CL
computes the delta between the two, and adds it to the total reported
size at the end of the request.

BUG=551000

Review-Url: https://codereview.chromium.org/2635433004
Cr-Original-Commit-Position: refs/heads/master@{#444753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a09d30c192709a7fcfdb97e9052f6f87e582b80