Implements CSSPropertyAPI for the scale property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyScale from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
scale property is stored in a CSSPropertyDescriptor, and is
called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIScale.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIScale.h is generated.
- Moves the parsing logic for scale from CSSPropertyParser.cpp to
  CSSPropertyAPIScale.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2639303005
Cr-Original-Commit-Position: refs/heads/master@{#444827}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f516b0e5891dac753279e7abb9b001db7c2c12f