Implements CSSPropertyAPI for the shape-outside property.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyShapeOutside from the switch statement in parseSingleValue,
and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
page property is stored in a CSSPropertyDescriptor, and is called from
CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIShapeOutside.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIShapeOutside.h is generated.
- Moves the parsing logic for page from CSSPropertyParser.cpp to
  CSSPropertyAPIShapeOutside.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2642013002
Cr-Original-Commit-Position: refs/heads/master@{#444958}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 81764f7d4806e67c5db9b46de3cf861a43af380b