Implements CSSPropertyAPI for the padding-{top,right,bottom,left} properties.
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyPaddingTop, CSSPropertyPaddingBottom, CSSPropertyPaddingLeft
and CSSPropertyPaddingRight from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
padding properties is stored in a CSSPropertyDescriptor, and is called
from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIPadding.cpp to the BUILD.gn file.
- Adds api_class=CSSPropertyAPIPadding flag to each border class property
  in CSSProperties.in, which indicates that CSSPropertyAPIPadding.h is
  generated.
- Moves the parsing logic for padding from CSSPropertyParser.cpp to
  CSSPropertyAPIPadding.cpp, which implements CSSPropertyAPI.h.

Review-Url: https://codereview.chromium.org/2645763003
Cr-Original-Commit-Position: refs/heads/master@{#444995}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e2d213e823ef2169f47aa88f55c04231ba124706