Crash in blink::beforeCallEnteredCallback()
Crash occurs because the script is forbidden due to the insertion
of the media element node when the media element is moved to a new document
and the existing play promise gets rejected because of the invoked load.

Check if the script is forbidden to skip reject play promises.

BUG=676004

Review-Url: https://codereview.chromium.org/2594353003
Cr-Original-Commit-Position: refs/heads/master@{#445107}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87800a0c858e3ee1e2a406b3aae00e89ae7d077f