Move PresentationAttributeCacheCleaner timer to main thread's timer task runner.
This is shared by multiple frames in the renderer main thread. Conceptually
it's UnspecedTimer, but since it's not associated with any particular frame,
and we know it will be used on the main thread, we can just use the main
thread's timer task queue.

BUG=624694

Review-Url: https://codereview.chromium.org/2647703006
Cr-Original-Commit-Position: refs/heads/master@{#445277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ca861148c5375e15b85831ee2826a818d682c93