Paint the whole of a table-part background image behind the cells in a table
Fix the painting of background images on table parts so that the image is displayed
across the whole of the part. This fixes each of table-row-group, table-row,
table-column and table-column-group.

The fix is pretty simple in principle: just adjust the geometry we use to paint the image
to reflect the cell's position relative to the table part from which the image originates.

Previously, we were just painting the image as though it was originating in each
cell. The one case left over is backgr_fixed-bg.html: this is still wrong and
I need to address it in a follow-up.

BUG=35697
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2600903002
Cr-Original-Commit-Position: refs/heads/master@{#445283}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 949b36db0565e9a81fc50cbb1584d64d6090c147