Implements CSSPropertyAPI for the CSSPropertyWebkitFontSizeDelta property.
A part of Project Ribbon, separating the parsing logic
for CSS properties from the parser into an API. This
patch removes CSSPropertyWebkitFontSizeDelta from the
switch statement in parseSingleValue, and calls the API
instead.

A function pointer to the parseSingleValue function from
the API for the CSSPropertyWebkitFontSizeDelta property
is stored in a CSSPropertyDescriptor, and is called from
CSSPropertyParser.

This patch:
- Adds CSSPropertyWebkitFontSizeDelta.cpp to the BUILD.gn
  file.
- Adds api_class flag to CSSProperties.in, which indicates
  that CSSPropertyWebkitFontSizeDelta.h is generated.
- Moves the parsing logic for CSSPropertyWebkitFontSizeDelta
  from CSSPropertyParser.cpp to
  CSSPropertyWebkitFontSizeDelta.cpp, which implements
  CSSPropertyAPI.h.

BUG=668012

Review-Url: https://codereview.chromium.org/2647503003
Cr-Original-Commit-Position: refs/heads/master@{#445327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 206f85553b9078dceb144cff4c4becb13f73624a