Implements CSSPropertyAPI for the column-span property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyColumnSpan from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
column-span property is stored in a CSSPropertyDescriptor,
and is called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIColumnSpan.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIColumnSpan.h is generated.
- Moves the parsing logic for column-span from
  CSSPropertyParser.cpp to CSSPropertyAPIColumnSpan.cpp,
  which implements CSSPropertyAPI.h.

BUG=668012
R=aazzam@google.com

Review-Url: https://codereview.chromium.org/2642823009
Cr-Original-Commit-Position: refs/heads/master@{#445335}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ddcd7967d46aa3fab2912723b59689d046b0b80d