Implements CSSPropertyAPI for the column-rule-width property
A part of Project Ribbon, separating the parsing logic for CSS
properties from the parser into an API. This patch removes
CSSPropertyColumnRuleWidth from the switch statement in
parseSingleValue, and calls the API instead.

A function pointer to the parseSingleValue function from the API for the
column-rule-width property is stored in a CSSPropertyDescriptor,
and is called from CSSPropertyParser.

This patch:
- Adds CSSPropertyAPIColumnRuleWidth.cpp to the BUILD.gn file.
- Adds api_class flag to CSSProperties.in, which indicates that
  CSSPropertyAPIColumnRuleWidth.h is generated.
- Moves the parsing logic for column-rule-width from
  CSSPropertyParser.cpp to CSSPropertyAPIColumnRuleWidth.cpp,
  which implements CSSPropertyAPI.h.

BUG=668012
R=aazzam@google.com

Review-Url: https://codereview.chromium.org/2641183004
Cr-Original-Commit-Position: refs/heads/master@{#445356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e59fefc829cee759865d3fd3e97ac4847c222824