Remove AudioSourceNode class
The MediaElementAudioSourceNode and MediaStreamAudioSourceNode were
implemented as deriving from an AudioSourceNode.  This is not part of
the spec; these nodes are derived from AudioNode.  This was only
visible in the prototype chain for these classes.  The AudioSourceNode
is otherwise invisible to the user; it is not constructible and has no
properties.

Thus, remove the AudioSourceNode class.

Intent: https://groups.google.com/a/chromium.org/d/msg/blink-dev/D-QJm9GCisc/gSCZwIrfFgAJ

BUG=663818
TEST=none

Review-Url: https://codereview.chromium.org/2621923003
Cr-Original-Commit-Position: refs/heads/master@{#445433}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 717012126c57dfe596b6562f9ad17f2a7e537a5b