Refactor ImageBitmap to stop drawing unpremul SkImage on SkCanvas
Skia is deprecating the use of Unpremultiplied alpha with SkImage. SkImage will still be able to store Unpremultiplied pixel values, but it will not do the right thing for drawing to an SkCanvas. ImageBitmap must be refactored to function properly after this change.

BUG=683154

Review-Url: https://codereview.chromium.org/2641393002
Cr-Original-Commit-Position: refs/heads/master@{#445495}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 12ec50eea00c7622356d6348a88f26e140a44d02